refactor: rename formatTime.ts to format-time.ts (kebab-case)
