+++ apps/rtk-query/src/player/playerSelectors.ts
@@ -3,7 +3,7 @@ import { createSelector } from '@reduxjs/toolkit'
 import type { RootState } from '@/app/store/store'
 
 import type { FormattedTime, TrackPlaybackState, TrackProgress } from './types/player.types'
-import { formatTime } from './utils/formatTime'
+import { formatTime } from './utils/format-time.ts'
 
 // ========================================
 // Basic Selectors
