+++ apps/rtk-query/src/player/utils/index.ts
@@ -1,3 +1,3 @@
-export { formatTime, parseTime } from './formatTime'
+export { formatTime, parseTime } from './format-time.ts'
 export { shuffle, shuffleWithCurrentItem } from './shuffle'
 export { debounce, throttle } from './throttle'
